Moved listener in the extension, lazy initialization
