+++ kie-server-parent/kie-server-services/kie-server-services-dmn/pom.xml
@@ -40,6 +40,10 @@
       <groupId>org.kie.server</groupId>
       <artifactId>kie-server-services-common</artifactId>
     </dependency>
+    <dependency>
+      <groupId>org.kie.server</groupId>
+      <artifactId>kie-server-services-prometheus</artifactId>
+    </dependency>
     
     <dependency>
       <groupId>org.kie</groupId>
