+++ kie-server-parent/kie-server-services/kie-server-services-prometheus/src/main/java/org/kie/server/services/prometheus/PrometheusMetricsListener.java
@@ -1,4 +1,4 @@
-package org.kie.server.services.dmn.metrics;
+package org.kie.server.services.prometheus;
 
 import io.prometheus.client.Histogram;
 import org.kie.dmn.api.core.event.AfterEvaluateBKMEvent;
